Front end work for Entry

+++ budgetentry/views.py
@@ -12,9 +12,10 @@ def homePageView(request):
 
 
 def viewDashboard(request):
-    uid = 1
-
-    request.session["uid"] = uid
+    if not request.user.is_authenticated():
+        return render(request, "/home")
+    uid = request.user.id
+    # request.session["uid"] = uid
 
     # cursor = connection.cursor()
     # cursor.execute("SET search_path TO postgres,public")
@@ -29,11 +30,10 @@ def viewDashboard(request):
     return render(request, 'budgetentry/dashboard.html', result)
 
 
-def viewEntries(request):
-    typeID = 1
-
-    request.session['typeID'] = typeID
-    uid = request.session['uid']
+def viewEntries(request, type_ID):
+    request.session['typeID'] = type_ID
+    uid = request.user.id
+    # uid = request.session['uid']
 
     # cursor = connection.cursor()
     # cursor.execute("SET search_path TO postgres,public")
@@ -41,29 +41,30 @@ def viewEntries(request):
 
     # entryList = [1, 2]  # BudgetEntryData
 
-    entryList = BudgetEntry.objects.filter(UID=uid, catID=typeID)
+    entryList = BudgetEntry.objects.filter(UID=uid, catID=type_ID)
 
-    result = {"entryList": entryList}
+    result = {"entryList": entryList, "typeID": type_ID}
     return render(request, "budgetentry/entrylist.html", result)
 
 
-def viewEntryDetail(request):
-    entryID = 1
+def viewEntryDetail(request, type_ID, entryID):
+    uid = request.user.id
+    # uid = request.session['uid']
+    # typeID = request.session['typeID']
 
-    uid = request.session['uid']
-    typeID = request.session['typeID']
-
-    data = BudgetEntry.objects.get(ID=entryID, UID=uid, catID=typeID)
-    result = {"entryID": entryID, "data": data, "type": BudgetType.objects.get(userID=uid, budgetTypeID=typeID)}
+    data = BudgetEntry.objects.get(ID=entryID, UID=uid, catID=type_ID)
+    result = {"entryID": entryID, "data": data, "type": BudgetType.objects.get(userID=uid, budgetTypeID=type_ID)}
     return render(request, "budgetentry/entrydetail.html", result)
 
 
-def createEntry(request):
+def createEntry(request, type_ID):
     if request.method == 'POST':
         form = CreateEntryForm(request.POST)
         if form.is_valid():
-            uid = request.session['uid']
-            catID = request.session['typeID']
+            # uid = request.session['uid']
+            # catID = request.session['typeID']
+            uid = request.user.id
+            catID = type_ID
             name = form.cleaned_data['name']
             date = form.cleaned_data['setDate']
             targetValue = form.cleaned_data['balance']
@@ -78,31 +79,35 @@ def createEntry(request):
     return render(request, "budgetentry/entrycrud.html", target)
 
 
-def editEntry(request):
-    entryID = 1
-
-    uid = request.session['uid']
-    typeID = request.session['typeID']
-
+def editEntry(request, type_ID, entryID):
+    uid = request.user.id
+    # entryID = 1
+    # uid = request.session['uid']
+    # typeID = request.session['typeID']
+    data = BudgetEntry.objects.get(ID=entryID, catID=type_ID)
+    if data.UID != uid:
+        return redirect("entries/")
     if request.method == 'POST':
         form = EditEntryForm(request.POST)
         if form.is_valid():
-            Entry = BudgetEntry.objects.get(ID=entryID, UID=uid, catID=typeID)
+            Entry = BudgetEntry.objects.get(ID=entryID, UID=uid, catID=type_ID)
             Entry.targetValue = form.cleaned_data['balance']
             Entry.save()
 
     mode = "edit"
-    data = BudgetEntry.objects.get(ID=entryID)
+
     target = {"mode": mode, "entryID": entryID, "data": data}
     return render(request, "budgetentry/entrycrud.html", target)
 
 
-def deleteEntry(request):  # Post
-    entryID = 1
-    Entry = BudgetEntry.objects.get(ID=entryID)
-    Entry.delete()
+def deleteEntry(request, type_ID, entryID):  # Post
+    # entryID = 1
+    if request.user.is_authenticated():
+        uid = request.user.id
+        Entry = BudgetEntry.objects.get(ID=entryID, catID=type_ID)
+        if Entry.UID == uid:
+            Entry.delete()
     return redirect("entries/")
-
 
 # def getEntry(entryID):
 #     # cursor = connection.cursor()
